refactor(store): extract session-history slice from appStore (part of #2077)
Move the recorded-session-history domain out of the appStore.ts monolith
into its own behavior-preserving Zustand slice, following the established
tunnel/embedded-servers/macros/plugins slice pattern.

The slice owns `sessionHistory` plus the load/record/pin/promote/remove/clear
actions and the private `stripHistorySecrets` helper. Actions still receive the
shared set/get typed against the full AppState, so the public store API and all
call sites (`useAppStore((s) => s.sessionHistory)`, `get().recordSession`, ...)
are unchanged.

Part of #2077
